Use Spring dependency injection for extensions.
The app project simply sets up Spring and JavaFX (and JavaFx Weaver)
as the platform.

The primary scene controller is extracted from the application.
This allows it to export plugins that other components can depend on
without creating a dependency cycle.

The original workspace componponents are separated into model and gui
components, with the gui component depending on and providing an extension
to the scene controller.